LS: Refactor `FileDiagnostics`
1. Moved this struct inside the ` diagnostics ` module.
2. Moved collecting and generating LSP messages as methods there.
3. This leaves `refresh_file_diagnostics` just taking care of state
   management and calling newly introduced functions.

commit-id:d838c83f
