+++ crates/cairo-lang-language-server/src/state.rs
@@ -2,10 +2,6 @@ use std::collections::HashSet;
 use std::ops::{Deref, DerefMut};
 use std::sync::Arc;
 
-use cairo_lang_diagnostics::Diagnostics;
-use cairo_lang_lowering::diagnostic::LoweringDiagnostic;
-use cairo_lang_parser::ParserDiagnostic;
-use cairo_lang_semantic::SemanticDiagnostic;
 use lsp_types::{ClientCapabilities, Url};
 use salsa::ParallelDatabase;
 
@@ -28,20 +24,6 @@ pub struct State {
     pub tricks: Owned<Tricks>,
     pub diagnostics_controller: DiagnosticsController,
 }
-
-#[derive(Clone, Default, PartialEq, Eq)]
-pub struct FileDiagnostics {
-    pub parser: Diagnostics<ParserDiagnostic>,
-    pub semantic: Diagnostics<SemanticDiagnostic>,
-    pub lowering: Diagnostics<LoweringDiagnostic>,
-}
-
-impl FileDiagnostics {
-    pub fn is_empty(&self) -> bool {
-        self.semantic.is_empty() && self.lowering.is_empty() && self.parser.is_empty()
-    }
-}
-impl std::panic::UnwindSafe for FileDiagnostics {}
 
 impl State {
     pub fn new(
